requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-jose>=3.3.0
python-multipart>=0.0.6
passlib>=1.7.4

